Authentication and Autherization are done

+++ controllers/resultsController.go
@@ -13,13 +13,27 @@ import (
 
 type ResultsController struct {
 	resultsService *services.ResultsService
+	usersService   *services.UsersService
 }
 
-func NewResultsController(resultsService *services.ResultsService) *ResultsController {
-	return &ResultsController{resultsService: resultsService}
+func NewResultsController(resultsService *services.ResultsService, usersService *services.UsersService) *ResultsController {
+	return &ResultsController{resultsService: resultsService, usersService: usersService}
 }
 
 func (rh ResultsController) CreateResult(ctx *gin.Context) {
+
+	// Authentication and Authorization
+	accessToken := ctx.Request.Header.Get("Token")
+	auth, resErr := rh.usersService.AuthorizeUser(accessToken, []string{ROLE_ADMIN})
+	if resErr != nil {
+		ctx.JSON(resErr.Status, resErr)
+		return
+	}
+	if !auth {
+		ctx.Status(http.StatusUnauthorized)
+		return
+	}
+
 	body, err := io.ReadAll(ctx.Request.Body)
 	if err != nil {
 		log.Println("Error while reading create result request body", err)
@@ -45,6 +59,19 @@ func (rh ResultsController) CreateResult(ctx *gin.Context) {
 }
 
 func (rh ResultsController) DeleteResult(ctx *gin.Context) {
+
+	// Authentication and Authorization
+	accessToken := ctx.Request.Header.Get("Token")
+	auth, resErr := rh.usersService.AuthorizeUser(accessToken, []string{ROLE_ADMIN})
+	if resErr != nil {
+		ctx.JSON(resErr.Status, resErr)
+		return
+	}
+	if !auth {
+		ctx.Status(http.StatusUnauthorized)
+		return
+	}
+
 	resultId := ctx.Param("id")
 	responseErr := rh.resultsService.DeleteResult(resultId)
 	if responseErr != nil {
